+++ src/main/java/net/neganote/monilabs/mixin/gt/IDurabilityBarMixin.java
@@ -1,4 +1,4 @@
-package net.neganote.monilabs.mixin;
+package net.neganote.monilabs.mixin.gt;
 
 import com.gregtechceu.gtceu.api.item.component.IDurabilityBar;
 import com.gregtechceu.gtceu.client.renderer.item.ToolChargeBarRenderer;
