+++ src/main/java/net/neganote/monilabs/mixin/gt/IMaterialPartItemMixin.java
@@ -1,4 +1,4 @@
-package net.neganote.monilabs.mixin;
+package net.neganote.monilabs.mixin.gt;
 
 import com.gregtechceu.gtceu.api.data.chemical.material.Material;
 import com.gregtechceu.gtceu.api.item.component.IMaterialPartItem;
