+++ src/main/java/net/neganote/monilabs/mixin/gt/LargeTurbineMachineMixin.java
@@ -1,4 +1,4 @@
-package net.neganote.monilabs.mixin;
+package net.neganote.monilabs.mixin.gt;
 
 import com.gregtechceu.gtceu.common.machine.multiblock.generator.LargeTurbineMachine;
