+++ src/main/resources/monilabs.mixins.json
@@ -7,14 +7,14 @@
   "refmap": "mixins.monilabs.refmap.json",
   "mixins": [
     "gt.DataAccessHatchMixin",
-    "IMaterialPartItemMixin",
+    "gt.IMaterialPartItemMixin",
     "KubeJSPluginsMixin",
     "gt.NotifiableEnergyContainerMixin",
     "gt.OpticalDataHatchMixin",
     "gt.PowerSubstationMachineMixin",
-    "LargeTurbineMachineMixin",
-    "RotorHolderMachineMixin",
-    "TurbineRotorBehaviourMixin",
+    "gt.LargeTurbineMachineMixin",
+    "gt.RotorHolderMachineMixin",
+    "gt.TurbineRotorBehaviourMixin",
     "aae.MixinCraftingJobStatusPacketCtorShim",
     "aae.MixinCraftingServiceAAE",
     "aae.MixinEncodedPatternItemShim",
@@ -32,7 +32,7 @@
     "microverses.ShaderLoaderMixin",
     "microverses.GameRendererMixin",
     "microverses.ProgramMixin",
-    "IDurabilityBarMixin",
+    "gt.IDurabilityBarMixin",
     "accessor.TerrainRenderPassAccessor",
     "accessor.ShaderChunkRendererAccessor",
     "microverses.IrisVertexFormatsMixin",
